requirements.txt
appnope==0.1.3
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
bcrypt==4.0.0
blinker==1.5
boto3==1.24.72
botocore==1.27.72
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
executing==1.0.0
flake8==5.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
idna==3.3
ipython==8.5.0
is-disposable-email==1.0.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.0
toml==0.10.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
wtforms-validators==1.0.0
